fix(pr): refactor a bit
